groweverythingserver/requirements.txt
alabaster==0.7.13
appdirs==1.4.4
asgiref==3.6.0
attrs==22.2.0
Babel==2.11.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
docutils==0.19
drf-spectacular==0.25.1
drf-yasg==1.21.4
esbonio==0.15.0
exceptiongroup==1.1.0
gunicorn==20.1.0
idna==3.4
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.24.1
packaging==23.0
pandas==1.5.2
pathspec==0.10.3
platformdirs==2.6.2
pluggy==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.4
pygls==0.13.1
Pygments==2.14.0
pyrsistent==0.19.3
pyspellchecker==0.7.1
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib.applehelp==1.0.3
sqlparse==0.4.3
tomli==2.0.1
typeguard==2.13.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
